refactor kubeadm join command generation
